Added functionality to catch if server is present or not

+++ python/deleteinstance.py
@@ -9,7 +9,17 @@ name = sys.argv[1]
 creds = get_nova_creds_v2()
 nova = Client(**creds)
  
-server = nova.servers.find(name=name)
-server.delete()
+servers_list = nova.servers.list()
+server_exists = False
 
-print(nova.servers.list())
+for s in servers_list:
+	if s.name == name:
+		server_exists = True
+		break
+if not server_exists:
+	print("The server %s does not exist." % name)
+else:
+	print("Deleting server: %s" % name)
+	server = nova.servers.find(name=name)
+	server.delete()
+	print("The server %s was deleted." % name)
